Fix all Flutter analyzer warnings: remove dead code and type errors
- ntfy_service.dart: fix type mismatch — asyncExpand<String> + typed onError callback
- home_page.dart: remove unused _showIpDialog (moved to SettingsScreen),
  remove unused _sessionDuration/_sessionStart, remove unnecessary ! operators
  on nullable vars where null check already done via ok == x != null
- fall_detector.dart: remove redundant cast (output[0][0] already double)

flutter analyze now reports 0 errors, 0 warnings (except deprecated withOpacity
in legacy activity_screen.dart which is not used by the main app flow).

+++ app/lib/inference/fall_detector.dart
@@ -120,7 +120,7 @@ class FallDetector {
       final input = [window.map((s) => s.map((v) => v.toDouble()).toList()).toList()];
       final output = [[0.0]];
       _interpreter!.run(input, output);
-      return (output[0][0] as double).clamp(0.0, 1.0);
+      return output[0][0].clamp(0.0, 1.0);
     }
 
     // Quantize float → int8
